Renamed files and folders for better naming convention
