+++ inventree_bulk_plugin/BulkGenerator/dimensions.py
