+++ inventree_bulk_plugin/BulkGenerator/generators/__init__.py
