+++ inventree_bulk_plugin/BulkGenerator/generators/generator.py
