+++ inventree_bulk_plugin/static/inventree-bulk/htm.LICENSE
