+++ inventree_bulk_plugin/static/inventree-bulk/preact.LICENSE
