+++ inventree_bulk_plugin/templates/components/BulkDefinitionSchemaBuilder.js
