+++ inventree_bulk_plugin/templates/components/constants.js
