+++ inventree_bulk_plugin/templates/components/generateBulkCreateView.js
