+++ inventree_bulk_plugin/templates/panels/category-detail/create-bulk.js
