+++ inventree_bulk_plugin/templates/panels/stock-index/manage-bulk.js
