+++ inventree_bulk_plugin/templates/panels/stock-location-detail/create-bulk.html
@@ -1,6 +1,6 @@
 {% load static %}
-<script src="{% static 'inventree-bulk/preact.js' %}"></script>
-<script src="{% static 'inventree-bulk/htm.js' %}"></script>
+<script src="{% static 'inventree-bulk-plugin/preact.js' %}"></script>
+<script src="{% static 'inventree-bulk-plugin/htm.js' %}"></script>
 
 <div id="bulk-create-location-preact-root"></div>
