+++ inventree_bulk_plugin/templates/panels/stock-location-detail/create-bulk.js
