+++ inventree_bulk_plugin/tests/unit/test_generators.py
